requirements.txt
filetype~=1.0.13
Pillow~=8.4.0
PySide6~=6.3.1